Fixed problem with MySQL not starting properly.
Installation was failing saying that MySQL couldn't install. Might just need to reboot automatically once the machine is provisioned and everything is downloaded.

MARS agent installation fails currently with System.Runtime.InteropServices.COMException trying to start a service, but it doesn't say which service. Will need to get to the source code when I get to work as it didn't show up in procmon what it was doing exactly.