Added custom script extension.
Ansible needs WinRM on to work, and Ansible supplies a prebuilt script for doing this in dev/test environments. So, the CSE downloads and executes the script. Confirmed that it works too.

Thoughts are that I can use CSEs for other things. Looks like you can only have one CSE or something at a time, but could have it download mobility service and execute the install process. Alternatively for the Linux replicated machine, could have it turn on the required root account for the push install. One or the other.